Overwrite old FEEC Maxwell tests
